1.4.5: Add reverse encoders for packed repeated fields. (#108)
When encoding from a Data.Vector.Vector this library
would already encode from right to left in order to
avoid saving all but the final element to the stack.
The relevant encoders would also exploit the fact
that vectors have known element counts in order to
consolidate size checks and associated reallocations
when encoding packed fields with fixed-width elements.

This change adds new encoders that generalize those
two optimizations to other types of container such as
`Data.Sequence.Seq`.  When a container does not support
right-to-left iteration and/or length prediction
(for example, a standard list), these new encoders
still support them as efficiently as is practical,
given those restrictions.

Also added:
  * bytesIfNonempty
  * etaMessageBuilder
  * repeatedMessageBuilder
  * unsafeFromByteString
  * unsafeFromShortByteString